Update legal and releases links
The new ember-website app uses slightly different URLs for the legal and releases pages. The app performs some URL translations (in the application controller) to convert the old legal and releases URLs to the new ones. This edit will update the old links so that the URL translations in the app will no longer be required and thus can be removed in the future.