Add link to Mastodon account
This adds a social media link to our new mastodon account.

It also adds `rel="me"` to the social media links, which is a semantic hint that these are pages about EmberJS on other sites, and which should cause our Mastodon profile to show as "Verified" as the owner of emberjs.com.